moving popper hook logic to its own file
